Add navigation mode overlays to GSI targets
This adds the navigation more overlays for three button and gesture
navigation to the GSI targets. This is done through
frameworks-base-overlays to maintain consistency with base targets.

Previously, the navigation mode overlay was missing, and the system
defaulted to three button navigation, despite missing the config values.
This lead to issues like an empty navigation mode settings page.

To maintain consistency with the previous GSI behaviour, this adds an
override to correctly set the three button navigation mode, using the
overlay. Note, base targets default to gesture navigation mode.

Flag: EXEMPT bugfix
Bug: 384552117
Test: flash GSI and verify OverlayManagerService state
  atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
Change-Id: I0ebb75f766ad5c0e32c7f19a210b75428e2733f5